refactor: remove SaveFileParams, inline save path, delete saveFile.ts
