+++ src/ElectronBackend/main/__tests__/get-save-file-listener.test.ts
@@ -82,10 +82,7 @@ describe('saveFileListener', () => {
 
     await saveFileListener(mainWindow)(AllowedFrontendChannels.SaveFileRequest);
 
-    expect(mockSaveFile).toHaveBeenCalledWith({
-      projectId: 'uuid_1',
-      opossumFilePath: '/my/file.opossum',
-    });
+    expect(mockSaveFile).toHaveBeenCalledWith('uuid_1', '/my/file.opossum');
     expect(dialog.showMessageBox).not.toHaveBeenCalled();
   });
